Move the `PrimaryButton` component to the `component` package.
