+++ client/src/main/kotlin/io/spine/examples/shareaware/client/component/Popup.kt
@@ -44,7 +44,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import io.spine.examples.shareaware.client.PrimaryButton
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
